Extract the notification sweep out of submissions.ts
Moves POST /api/internal/notify-sweep into
notifications/notify-sweep-routes.ts as registerNotifySweepRoutes(app,
deps), with HANDOFF_ACK_STALL_MS -- whose only reference in the whole
repo was this route. submissions.ts drops 3723 -> 3479; the new file is
273 lines against the 500 cap.

The bucket is notifications rather than creation because the route's two
main collaborators, notifyOnTransition and detectOperatorAlerts, already
live there, so most of its imports become same-bucket edges.

Fifteen seams, passed by reference in the established way. A stale
scoping note claimed fourteen and named five that do not appear in the
route at all -- stagedPreviews among them, which is declared *after* the
extraction point and could never have been captured. Wiring that list as
given would not have compiled. recordDerivedJobState stays a dep rather
than moving with the route: it closes over store, now and app.log, and
its JSDoc is inside submissions.ts's frozen prose baseline.

Six imports went unused in submissions.ts once the route left and are
dropped, or --max-warnings 0 would have failed on them.

Proved the move is behaviour-preserving by comparing the two bodies with
comments and blank lines stripped: 183 code lines each, identical. Only
comments differ, rewritten as one-liners because a new file's prose
baseline starts at zero -- verified with the real findCommentProseViolations
rather than by counting words, which the checker had already caught me
getting wrong twice on hyphens and on a bare // failing to break a stack.

Two pre-existing inaccuracies surfaced while compressing. The lead
comment claimed a 30/hour rate ceiling where the code says 120, so the
replacement states no number at all. And the stall log line still blamed
"the games-repo @copilot relay", a path removed months ago -- it now
says no agent has collected the feedback, which is what actually
happened. An operator reading the old line mid-incident would have gone
looking for a subsystem that no longer exists.

Both ratchets lowered to the new sizes so the gain cannot quietly erode.
